feat: refactor ClassDetails component and add new Schedule and Activities features
