+++ pes/server/pes_be/src/main/java/com/sba301/group1/pes_be/PesBeApplication.java
@@ -58,6 +58,10 @@ public class PesBeApplication {
 
     private final StudentClassRepo studentClassRepo;
 
+    private final StudentRepo studentRepo;
+
+    private final ParentRepo parentRepo;
+
     private final EntityManager entityManager;
 
     public static void main(String[] args) {
