+++ pes/server/pes_be/src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.application.name=pes_be
 spring.datasource.url=jdbc:sqlserver://localhost:1433;databaseName=pes_sba;encrypt=true;trustServerCertificate=true
 
 spring.datasource.username=sa
-spring.datasource.password=12345
+spring.datasource.password=Password123!
 spring.datasource.driver-class-name=com.microsoft.sqlserver.jdbc.SQLServerDriver
 
 spring.jpa.database=SQL_SERVER
